+++ sdk/golang/pkg/generate/margin/market/types_get_mark_price_detail_resp.go
@@ -15,12 +15,12 @@ type GetMarkPriceDetailResp struct {
 	// Timestamp (milliseconds)
 	TimePoint int64 `json:"timePoint,omitempty"`
 	// Mark price
-	Value float32 `json:"value,omitempty"`
+	Value float64 `json:"value,omitempty"`
 }
 
 // NewGetMarkPriceDetailResp instantiates a new GetMarkPriceDetailResp object
 // This constructor will assign default values to properties that have it defined
-func NewGetMarkPriceDetailResp(symbol string, timePoint int64, value float32) *GetMarkPriceDetailResp {
+func NewGetMarkPriceDetailResp(symbol string, timePoint int64, value float64) *GetMarkPriceDetailResp {
 	this := GetMarkPriceDetailResp{}
 	this.Symbol = symbol
 	this.TimePoint = timePoint
